Cleaned lots of scripts

+++ OpenGLGameEngine/OpenGLGameEngine/CameraMovement.cpp
@@ -33,7 +33,7 @@ void CameraMovement::Update(double deltaTime) {
 		float horizontalMoveSpeed = lookSpeed * (float)Input::getAxis(InputAxisMouseX);
 		transform->rotate(glm::vec3(0, horizontalMoveSpeed, 0));
 
-		float verticalMoveSpeed = lookSpeed * Input::getAxis(InputAxisMouseY);
+		float verticalMoveSpeed = lookSpeed * (float)Input::getAxis(InputAxisMouseY);
 		transform->rotate(glm::vec3(verticalMoveSpeed, 0, verticalMoveSpeed) * (transform->rotation * glm::vec3(0, 0, 1)));
 	}
